+++ northstar_tests/src/runtime.rs
@@ -12,38 +12,38 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-//! Controls North runtime instances
+//! Controls Northstar runtime instances
 
 use crate::process_assert::ProcessAssert;
 use color_eyre::eyre::{eyre, Error, Result, WrapErr};
-use north::{api, runtime, runtime::config::Config};
+use northstar::{api, runtime, runtime::config::Config};
 use std::{fs, path::Path};
 use tokio::{time, time::timeout};
 
 const TIMEOUT: time::Duration = time::Duration::from_secs(3);
 
-/// A running instance of north.
+/// A running instance of northstar.
 pub struct Runtime(runtime::Runtime);
 
 impl Runtime {
-    /// Launches an instance of north
+    /// Launches an instance of northstar
     pub async fn launch() -> Result<Runtime, Error> {
         let config = toml::from_str::<Config>(
-            &fs::read_to_string("north.toml").wrap_err("Failed to read north.toml")?,
+            &fs::read_to_string("northstar.toml").wrap_err("Failed to read northstar.toml")?,
         )
-        .wrap_err("Failed to parse north.toml")?;
+        .wrap_err("Failed to parse northstar.toml")?;
         let runtime = timeout(TIMEOUT, runtime::Runtime::start(config))
             .await
-            .wrap_err("Launching north timed out")
-            .and_then(|result| result.wrap_err("Failed to instantiate north runtime"))?;
+            .wrap_err("Launching northstar timed out")
+            .and_then(|result| result.wrap_err("Failed to instantiate northstar runtime"))?;
         Ok(Runtime(runtime))
     }
 
     pub async fn launch_with_config(config: runtime::config::Config) -> Result<Runtime, Error> {
         let runtime = timeout(TIMEOUT, runtime::Runtime::start(config))
             .await
-            .wrap_err("Launching north timed out")
-            .and_then(|result| result.wrap_err("Failed to instantiate north runtime"))?;
+            .wrap_err("Launching northstar timed out")
+            .and_then(|result| result.wrap_err("Failed to instantiate northstar runtime"))?;
         Ok(Runtime(runtime))
     }